UI improve (#4)
* 为OIDC Client的管理更改路由约定。

* UI结构调整。

* 改进从枚举生成SelectListItem集合的方式。

* Make OperationResult obsolete.

* Refact using IdentityResult instead of OperationResult.

* Reuse OperationResult, but rename to IdOperationResult.

* Correct title localization with format pattern.

* Localization and some other changes.

* 优化视图资源，调整命名路径等。

* UI调整。

* 优化设计数据库迁移，为初始化工具做准备。
Squashed commit of the following:

commit 19595970c837066673e180dd1cf350a0ee78129e
Author: Jonny Yang <63300503@qq.com>
Date:   Mon Oct 30 02:25:47 2023 +0800

    重新设计DatabaseTool，为初始化工具做准备。

commit 45793d2b6c3bb6d27fd7817f391d16ba0c801aa8
Author: Jonny Yang <63300503@qq.com>
Date:   Sun Oct 29 20:11:06 2023 +0800

    Rename folder.

commit a6cf4f80f6127279c272f5c91b938e4f1c2723ec
Author: Jonny Yang <63300503@qq.com>
Date:   Sun Oct 29 19:31:32 2023 +0800

    nuget upgrade.

commit ddc3499f736dbf73d7e14b6705fd0d41e7467625
Author: Jonny Yang <63300503@qq.com>
Date:   Sun Oct 29 19:24:41 2023 +0800

    adjust namespaces.

commit a33b14ce49e61b6900db231a8bc6a3a30802b94d
Author: Jonny Yang <63300503@qq.com>
Date:   Sun Oct 29 19:17:47 2023 +0800

    rename projects.

commit 3fefc0e000e346e6fa00611948a0883b81ef2c52
Author: Jonny Yang <63300503@qq.com>
Date:   Sun Oct 29 19:13:59 2023 +0800

    将各个DbContext分开到单独程序集。清除过时的Migrations.

* 优化登录时错误消息布局。

* Rename view localizer to L for shorten spelling.
Add html localizer via shared resource if necessary.

* update testing sample data.

* Rename ProductInfo at view imports.

* UI improve for person.

* UI improve for people and organization.

* Code clean.

* 调整用户管理器逻辑。

* 改进注入支持。

* 解决Lazy loading时设置用户名的BUG.
清理项目并设置过滤方案。

* Remove iqueryable organization store, fix some issues.

* configure serilog.

* Code clean.

* Rename projects and adjust directory structures.

* repo clean.

* 调整RealName工程结构。

* 在AuthCenter上实现用户头像管理。

* 优化Profile picture逻辑。

* Localization.

* 初步迁移了账户管理功能。

* Move sex and take localization works.

* Remove my account area.

* 解除对Identity.Stores的耦合。

* 添加优化某些测试。

* 调整使用正确的日期时间类型。

* some corrections.

* 本地化。

* 更新文档。

* 一些更改。

* Code Clean.

* Code Clean.

* 添加密码过期和重设功能。

* 重制迁移脚本。

* 修正字段。调整设置。

* Upgrade referenced packages.

---------

+++ src/AlphaID/IDSubjects/OrganizationMemberManager.cs
@@ -9,10 +9,10 @@ namespace IDSubjects;
 public class OrganizationMemberManager
 {
     private readonly IOrganizationMemberStore store;
-    readonly ILogger<OrganizationMemberManager>? logger;
+    private readonly ILogger<OrganizationMemberManager>? logger;
 
     /// <summary>
-    /// Init GenericOrganization Member Mamager via GenericOrganization Member store.
+    /// Init GenericOrganization Member Manager via GenericOrganization Member store.
     /// </summary>
     /// <param name="store"></param>
     /// <param name="logger"></param>
@@ -45,7 +45,7 @@ public class OrganizationMemberManager
         var members = this.store.OrganizationMembers.Where(p => p.OrganizationId == organization.Id);
         Debug.Assert(members != null);
         var visibilityLevel = MembershipVisibility.Public;
-        if(visitor != null)
+        if (visitor != null)
         {
             visibilityLevel = MembershipVisibility.AuthenticatedUser;
             if (members.Any(m => m.PersonId == visitor.Id))
@@ -54,6 +54,11 @@ public class OrganizationMemberManager
         return Task.FromResult(members.Where(m => m.Visibility >= visibilityLevel).AsEnumerable());
     }
 
+    /// <summary>
+    /// Get organization members.
+    /// </summary>
+    /// <param name="organization">Organization</param>
+    /// <returns></returns>
     public Task<IEnumerable<OrganizationMember>> GetMembersAsync(GenericOrganization organization)
     {
         var members = this.store.OrganizationMembers.Where(p => p.OrganizationId == organization.Id);
@@ -71,15 +76,20 @@ public class OrganizationMemberManager
         var members = this.store.OrganizationMembers.Where(p => p.PersonId == person.Id);
         Debug.Assert(members != null);
         var visibilityLevel = MembershipVisibility.Public;
-        if(visitor != null)
+        if (visitor != null)
         {
             visibilityLevel = MembershipVisibility.AuthenticatedUser;
-            if(members.Any(m => m.PersonId == visitor.Id))
+            if (members.Any(m => m.PersonId == visitor.Id))
                 visibilityLevel = MembershipVisibility.Private; //Visitor is a member of the organization.
         }
         return Task.FromResult(members.Where(m => m.Visibility >= visibilityLevel).AsEnumerable());
     }
 
+    /// <summary>
+    /// 获取个人的组织成员身份。
+    /// </summary>
+    /// <param name="person"></param>
+    /// <returns></returns>
     public Task<IEnumerable<OrganizationMember>> GetMembersOfAsync(NaturalPerson person)
     {
         var members = this.store.OrganizationMembers.Where(p => p.PersonId == person.Id);
@@ -87,41 +97,16 @@ public class OrganizationMemberManager
     }
 
     /// <summary>
-    /// Take person join in organization.
-    /// </summary>
-    /// <param name="person"></param>
-    /// <param name="organization"></param>
-    /// <param name="title"></param>
-    /// <param name="department"></param>
-    /// <param name="remark"></param>
-    /// <returns></returns>
-    public async Task<OperationResult> JoinOrganizationAsync(NaturalPerson person, GenericOrganization organization, string? title = null, string? department = null, string? remark = null)
-    {
-        var member = await this.GetMemberAsync(person, organization);
-        if (member != null)
-            return OperationResult.Error("自然人已是该组织的成员。");
-
-        member = new(organization, person)
-        {
-            Title = title,
-            Department = department,
-            Remark = remark,
-        };
-        await this.store.CreateAsync(member);
-        return OperationResult.Success;
-    }
-
-    /// <summary>
     /// 
     /// </summary>
     /// <param name="member"></param>
     /// <returns></returns>
-    public async Task<OperationResult> CreateAsync(OrganizationMember member)
+    public async Task<IdOperationResult> CreateAsync(OrganizationMember member)
     {
         if (this.store.OrganizationMembers.Any(p => p.OrganizationId == member.OrganizationId && p.PersonId == member.PersonId))
-            return OperationResult.Error("Member already exists.");
+            return IdOperationResult.Failed(Resources.MembershipExists);
         await this.store.CreateAsync(member);
-        return OperationResult.Success;
+        return IdOperationResult.Success;
     }
 
     /// <summary>
@@ -130,14 +115,27 @@ public class OrganizationMemberManager
     /// <param name="person"></param>
     /// <param name="organization"></param>
     /// <returns></returns>
-    public async Task<OperationResult> LeaveOrganizationAsync(NaturalPerson person, GenericOrganization organization)
+    public async Task<IdOperationResult> LeaveOrganizationAsync(NaturalPerson person, GenericOrganization organization)
     {
-        var member = await this.GetMemberAsync(person, organization);
+        var members = await this.GetMembersAsync(organization);
+        var member = members.FirstOrDefault(m => m.PersonId == person.Id);
         if (member != null)
         {
-            await this.store.DeleteAsync(member);
+            if (member.IsOwner && members.Count(m => m.IsOwner) <= 1)
+                return IdOperationResult.Failed(Resources.LastOwnerCannotLeave);
+            return await this.store.DeleteAsync(member);
         }
-        return OperationResult.Success;
+        return IdOperationResult.Success;
+    }
+
+    /// <summary>
+    /// 移除用户成员身份，无论用户是否是组织所有者也是如此。
+    /// </summary>
+    /// <param name="member"></param>
+    /// <returns></returns>
+    public async Task<IdOperationResult> RemoveAsync(OrganizationMember member)
+    {
+        return await this.store.DeleteAsync(member);
     }
 
     /// <summary>
@@ -145,9 +143,8 @@ public class OrganizationMemberManager
     /// </summary>
     /// <param name="member"></param>
     /// <returns></returns>
-    public async Task<OperationResult> UpdateAsync(OrganizationMember member)
+    public async Task<IdOperationResult> UpdateAsync(OrganizationMember member)
     {
-        await this.store.UpdateAsync(member);
-        return OperationResult.Success;
+        return await this.store.UpdateAsync(member);
     }
 }
